Modify FETCH_USER, login/logout, dashboard, desktop container to work correctly with redux state
