+++ client/src/actions/types.js
@@ -17,10 +17,12 @@ export const FETCH_MANAGERS = 'FETCH_MANAGERS';
 // @ USER ACTION TYPES
 // @ @ @ @ @ @ @ @ @ @ @ @
 export const FETCH_USER = 'FETCH_USER';
+export const FETCH_USERS = 'FETCH_USERS';
 export const GET_MESSAGES = 'GET_MESSAGES';
 export const MESSAGE_USER = 'MESSAGE_USER';
 export const ADD_EMPLOYEE = 'ADD_EMPLOYEE';
 export const ADD_DEPARTMENT = 'ADD_DEPARTMENT';
+export const VERIFIED_USER = 'VERIFIED_USER';
 
 // @ @ @ @ @ @ @ @ @ @ @ @
 // @ SCHEDULES ACTION TYPES
